jwt devredışı bırakldı ve birçok güvenlik güncellemesi yapıldı

+++ webapideneme1/src/main/java/com/demo/webapideneme1/services/GroupService.java
@@ -1,5 +1,7 @@
 package com.demo.webapideneme1.services;
 
+import java.security.Principal;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -10,20 +12,25 @@ import org.springframework.transaction.annotation.Transactional;
 import com.demo.webapideneme1.models.Comment;
 import com.demo.webapideneme1.models.Group;
 import com.demo.webapideneme1.models.User;
+import com.demo.webapideneme1.repositories.CommentRepository;
 import com.demo.webapideneme1.repositories.GroupRepository;
+import com.demo.webapideneme1.repositories.UserRepository;
+
+import jakarta.servlet.http.HttpServletRequest;
 
 @Service
 public class GroupService {
 	GroupRepository groupRepository;
-	UserService userService;
-	CommentService commentService;
+	UserRepository userRepository;
+	CommentRepository commentRepository;
 	
 	@Autowired
-	public GroupService(GroupRepository groupRepository, UserService userService,CommentService commentService) {
+	public GroupService(GroupRepository groupRepository, 
+			UserRepository userRepository,CommentRepository commentRepository) {
 		super();
 		this.groupRepository = groupRepository;
-		this.userService = userService;
-		this.commentService=commentService;
+		this.userRepository = userRepository;
+		this.commentRepository=commentRepository;
 	}
 
 	public Group getOneGroupById(Long groupId) {
@@ -50,20 +57,31 @@ public class GroupService {
 		return groups;
 	}
 	@Transactional
-	public void deleteGroupById(Group group) {
-		List<Comment> comments=commentService.getCommentsOfAGroup(group.getId());
-		while(comments.size()>0)
+	public String deleteGroupById(HttpServletRequest request,long groupId) {
+		/*jwt olmadan requestten kullanıcı adını alma kodları başlangıcı*/		
+		Principal pl=request.getUserPrincipal();
+		String username=pl.getName();
+		/*jwt olmadan requestten kullanıcı adını alma kodları sonu*/
+		User user=userRepository.findByUsername(username);
+		Group group=groupRepository.findById(groupId).orElse(null);
+		if(group!=null&&(group.getOwner()==user||user.getRoles().contains("ADMIN")))
 		{
-			
-			comments.get(0).setGroup(null);
-			comments.get(0).setOwner(null);
-			comments.get(0).setQuotedComment(null);
-			commentService.saveComment(comments.get(0));
-			commentService.deleteComment(comments.get(0).getId());
-			comments.remove(comments.get(0));
-			
-		}
-		groupRepository.delete(group);
+			List<Comment> comments=commentRepository.findByGroupId(groupId);
+			while(comments.size()>0)
+			{
+				
+				comments.get(0).setGroup(null);
+				comments.get(0).setOwner(null);
+				comments.get(0).setQuotedComment(null);
+				commentRepository.save(comments.get(0));
+				commentRepository.delete(comments.get(0));
+				comments.remove(comments.get(0));
+				
+			}
+			groupRepository.delete(group);
+			return "success";
+		}else return "fail";
+		
 		
 	}
 
@@ -73,7 +91,7 @@ public class GroupService {
 	}
 
 	public List<Group> getGroupsOfAMember(long userId) {
-		User member=userService.getOneUserById(userId);
+		User member=userRepository.findById(userId).orElse(null);
 		if(member!=null)
 		{
 			if(member!=null)
@@ -84,6 +102,58 @@ public class GroupService {
 		}else
 		
 		return null;
+	}
+
+	public String createGroup(HttpServletRequest request, String name) {
+		/*jwt olmadan requestten kullanıcı adını alma kodları başlangıcı*/		
+		Principal pl=request.getUserPrincipal();
+		String username=pl.getName();
+		/*jwt olmadan requestten kullanıcı adını alma kodları sonu*/
+		User owner=userRepository.findByUsername(username);
+		if(!name.matches("^[öüÖÜĞğşŞçÇıİ|a-z|A-Z]{1,20}(\\s[öüÖÜĞğşŞçÇıİ|a-z|A-Z]{1,20}){0,10}$"))
+			return "Grup is not suitable to format. Group could not be created";
+		List<Group> groups=groupRepository.findAll();
+		for(Group g : groups)
+		{
+			if(g.getName().equals(name)) return "A group name cannot be the same as the name of another group. Group creation unsuccessful";
+		}
+		if(owner==null) return "Group cannot be created. Group owner not found";
+		if(owner!=null)
+		{
+			Group group=new Group();
+			group.setOwner(owner);
+			if(group.getMembers()!=null)
+			{
+				group.getMembers().add(owner);
+				group.setMembers(group.getMembers());
+			}else
+			{
+				List<User> members=new ArrayList<User>();
+				members.add(owner);
+				group.setMembers(members);
+			}
+			 
+			group.setName(name);
+			groupRepository.save(group);
+			return "Group successfully created";
+		
+		}else return "Group could not be created";
+	}
+
+	public String updateGroupName(HttpServletRequest request, long groupId, String newgroupname) {
+		/*jwt olmadan requestten kullanıcı adını alma kodları başlangıcı*/		
+		Principal pl=request.getUserPrincipal();
+		String username=pl.getName();
+		/*jwt olmadan requestten kullanıcı adını alma kodları sonu*/
+		User user=userRepository.findByUsername(username);
+		Group group=groupRepository.findById(groupId).orElse(null);
+		if(user!=null && group!=null && user==group.getOwner() )
+		{
+			group.setName(newgroupname);
+			groupRepository.save(group);
+			return "Group name successfully updated";
+		}
+		return "Group name cannot be updated";
 	}
